Moved task to public package
