+++ src/main/groovy/org/shipkit/internal/gradle/contributors/ContributorsPlugin.java
@@ -4,6 +4,7 @@ import org.gradle.api.Action;
 import org.gradle.api.Plugin;
 import org.gradle.api.Project;
 import org.shipkit.gradle.ReleaseConfiguration;
+import org.shipkit.gradle.notes.ContributorsFetcherTask;
 import org.shipkit.internal.gradle.configuration.ReleaseConfigurationPlugin;
 import org.shipkit.internal.gradle.util.TaskMaker;
 
@@ -20,7 +21,7 @@ import static org.shipkit.internal.gradle.util.BuildConventions.contributorsFile
  *
  * Adds tasks:
  * <ul>
- *     <li>fetchAllContributors - {@link AllContributorsFetcherTask}</li>
+ *     <li>fetchAllContributors - {@link ContributorsFetcherTask}</li>
  * </ul>
  */
 public class ContributorsPlugin implements Plugin<Project> {
@@ -33,9 +34,9 @@ public class ContributorsPlugin implements Plugin<Project> {
     }
 
     private void fetchAllTask(final Project project, final ReleaseConfiguration conf) {
-        project.getTasks().create(FETCH_ALL_CONTRIBUTORS_TASK, AllContributorsFetcherTask.class, new Action<AllContributorsFetcherTask>() {
+        project.getTasks().create(FETCH_ALL_CONTRIBUTORS_TASK, ContributorsFetcherTask.class, new Action<ContributorsFetcherTask>() {
             @Override
-            public void execute(final AllContributorsFetcherTask task) {
+            public void execute(final ContributorsFetcherTask task) {
                 task.setGroup(TaskMaker.TASK_GROUP);
                 task.setDescription("Fetch info about all project contributors from GitHub and store it in file");
                 task.setOutputFile(contributorsFile(project));
